+++ src/main/java/com/lsmman/mykarrotservice/product/application/ProductService.java
@@ -1,5 +1,7 @@
-package com.lsmman.mykarrotservice.product;
+package com.lsmman.mykarrotservice.product.application;
 
+import com.lsmman.mykarrotservice.product.domain.Product;
+import com.lsmman.mykarrotservice.product.domain.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
