+++ src/main/java/com/lsmman/mykarrotservice/product/domain/DealStatus.java
@@ -1,4 +1,4 @@
-package com.lsmman.mykarrotservice.product;
+package com.lsmman.mykarrotservice.product.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
