+++ src/main/java/com/lsmman/mykarrotservice/product/domain/ProductRepository.java
@@ -1,5 +1,6 @@
-package com.lsmman.mykarrotservice.product;
+package com.lsmman.mykarrotservice.product.domain;
 
+import com.lsmman.mykarrotservice.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
